api/requirements.txt
Flask
mysql-connector-python
pika
